requirements.txt
glfw==2.5.9
numpy==1.24.3
PyOpenGL==3.1.7
PyGLM
scipy
